requirements.txt
ccxt
python-telegram-bot==20.3
pandas
numpy
ta
